old/ai_tools/requirements.txt
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
discord.py>=2.3.0
anthropic>=0.40.0
